Fixed issues, improved Markdown Readme, added different fluid zones.
Added Mudzone/Icyzone prefabs that affect the player and other rigidbodies properly as they travel through it.
Fixed issue with enemy dying instantly when hit by the player.
Fixed issue preventing collision damage from happening on rigidbodies.
The in-game wiki-page/readme editor can now be edited in the inspector itself, including adding/changing big title and icon.
Made a custom inspector for goal so it will warn you if the scene name isn't set properly (or that scene name hasn't been added to the build info). It also has a button with a dropdown to set the scene name from one of the ones in the build info.
Coins now also get added to the screen in the bottom left.

Also made it so if there is no camera in the scene one will be spawned for you by default, and when there are no players it says on the screen Press A or Space to join.